requirements.txt
Flask==2.3.2
opencv-python-headless==4.8.1.78
Pillow==9.5.0
matplotlib==3.7.1
torch==2.0.1
ultralytics==8.0.154
gunicorn==20.1.0